html/browsers/browsing-the-web/navigating-across-documents/005.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [6]

description: |
  1. Creates a srs-less iframe with an `id` of `test` and `name` of `test`
  2. Creates an anchor element with three attrs:
    a. `target` attr set to `test`, making the iframe this link's target
    b. `onclick` handler that selects the iframe in the DOM and sets its `location` to `click.html`.
    c. `href` attr with a value of `href.html`
  3. href.thml contains script that postmessages back the string "href"
  4. click.html contains script that postmessages back the string "click"
  5. 005.html then contains an async test which
    a. clicks the Link
    b. listens for the post message response
  6. `onclick` handlers get evaluated before hrefs, since the href is second, it is brought to maturity, and onclick's location change should be aborted per step 6 of 7.8.1

/-->
<title>Link with onclick navigation is aboprted by href navigation </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.getElementById('test').contentWindow.location='click.html'" href="href.html">Test</a>
<script>
var test = async_test();
test.step(function() {
  document.querySelector("a").click();
});
window.onmessage = test.step_func(
  function(e) {
    assert_equals(e.data, "href");
    test.done();
  });
</script>
